refactor error handling to new errorr_handling file
