+++ ai_commit_msg/services/openai_service.py
@@ -1,9 +1,10 @@
 from logging import Logger
+from ai_commit_msg.utils.errorr_handling import map_error
 from openai import OpenAI
 from ai_commit_msg.services.config_service import ConfigService
 from ai_commit_msg.services.local_db_service import LocalDbService, CONFIG_COLLECTION_KEY
 from ai_commit_msg.utils.models import OPEN_AI_MODEL_LIST
-from ai_commit_msg.services.model_error_handling import map_error
+
 
 class OpenAiService:
     client = None
